Refactored MainLayout to use the Menu component. Combined SimpleForm and VariableForm into one page. Added "Remove variation" button.
